review changes

+++ tests/testutils.py
@@ -1,8 +1,17 @@
 import os
 import shutil
 import sys
+from typing import Iterable
 
-def shell_exec(cmd, exit_on_fail=True, echo=True) -> int:
+def init(caller_file: str = '', clean_dirs: Iterable[str] = []):
+    # Needed for mama commands to perform work in the correct directory
+    if caller_file:
+        os.chdir(os.path.dirname(os.path.abspath(caller_file)))
+
+    for d in clean_dirs:
+        rmdir(d)
+
+def shell_exec(cmd: str, exit_on_fail: bool = True, echo: bool = True) -> int:
     if echo: print(f'exec: {cmd}')
     result = os.system(cmd)
     if result != 0 and exit_on_fail:
@@ -12,18 +21,18 @@ def shell_exec(cmd, exit_on_fail=True, echo=True) -> int:
         sys.exit(result)
     return result
 
-def file_contains(filepath, text):
+def file_contains(filepath: str, text: str) -> bool:
     with open(filepath, 'r') as f:
         content = f.read()
     return text in content
 
-def file_exists(filepath):
+def file_exists(filepath: str) -> bool:
     return os.path.isfile(filepath)
 
-def is_windows():
+def is_windows() -> bool:
     return os.name == 'nt'
 
-def is_linux():
+def is_linux() -> bool:
     return os.name == 'posix' and sys.platform != 'darwin'
 
 def onerror(func, path, _):
@@ -32,5 +41,6 @@ def onerror(func, path, _):
         os.chmod(path, stat.S_IWUSR)
         func(path)
 
-def rmtree(path: str):
-    shutil.rmtree(path, onerror=onerror)
+def rmdir(path: str):
+    if os.path.exists(path):
+        shutil.rmtree(path, onerror=onerror)
